+++ src/templates/Candidate.js
@@ -9,7 +9,7 @@ import { BLOCKS } from '@contentful/rich-text-types'
 
 export default function CandidateTemplate({ pageContext }) {
   const { candidate } = pageContext
-  const { name, position, vision, image, platforms } = candidate
+  const { name, position, vision, image, platforms, portfolio } = candidate
   const options = {
     renderNode: {
       [BLOCKS.UL_LIST]: (node, children) => {
@@ -45,6 +45,33 @@ export default function CandidateTemplate({ pageContext }) {
             src={image.file.url}
             className={`h-full w-full object-cover object-top rounded-xl md:rounded-none`}
           />
+          {portfolio && (
+            <div className="absolute bottom-0 rounded-b-xl p-2 bg-[rgba(0,0,0,.5)] w-full text-white">
+              <a
+                href={portfolio.file.url}
+                target="_blank"
+                className="flex gap-2 justify-center"
+              >
+                <p>Portfolio</p>
+                <svg
+                  width="24"
+                  height="24"
+                  viewBox="0 0 24 24"
+                  fill="none"
+                  xmlns="http://www.w3.org/2000/svg"
+                >
+                  <path
+                    d="M15.6396 7.02527H12.0181V5.02527H19.0181V12.0253H17.0181V8.47528L12.1042 13.3892L10.6899 11.975L15.6396 7.02527Z"
+                    fill="currentColor"
+                  />
+                  <path
+                    d="M10.9819 6.97473H4.98193V18.9747H16.9819V12.9747H14.9819V16.9747H6.98193V8.97473H10.9819V6.97473Z"
+                    fill="currentColor"
+                  />
+                </svg>
+              </a>
+            </div>
+          )}
         </div>
         <div className="flex flex-col justify-end gap-8 py-8 p-0 md:p-10 lg:p-12 text-white">
           <div>